feat(firmware): port to Waveshare ESP32-C6-LCD-1.47
- Add BOARD_WAVESHARE_C6_LCD_1_47 to BoardConfig.h: cadence on GPIO 18,
  potentiometer on GPIO 1 (ADC1_CH1), calibration button on GPIO 9
  (onboard BOOT). Avoids GPIO 4/5 (TF card) and other onboard usage.
- New [env:waveshare_c6_lcd_1_47] in platformio.ini using pioarduino fork
  (espressif32 platform doesn't yet ship Arduino for C6). Uses ESP32Async
  AsyncTCP/ESPAsyncWebServer (esphome fork's IPv6Address.h is gone in
  Arduino-ESP32 3.x). Excludes ILI9341 / LCD1602 / touch sources via
  build_src_filter.
- St7789Ui: minimal IDisplay implementation using Arduino_GFX_Library for
  the 172x320 panel (column offset 34, BGR + inversion). Re-asserts
  backlight every tick and forces a periodic full repaint to defend
  against frozen frames.
- BleCps: fix NimBLE 2.x build (include Arduino.h + NimBLEAdvertising.h,
  start the server instead of each service individually).
- PowerReal: use 64-bit arithmetic for crank_evt_1024 so uptime > 70 min
  doesn't overflow uint32_t.
- wroom-monark.ino: enable display + per-second SAMPLE serial line (rpm,
  kp, power, adc, rev count, evt time, uptime, free heap) for live
  observability during dev rides.

+++ PowerReal.cpp
@@ -273,11 +273,10 @@ void PowerReal::update(uint32_t now_ms) {
   Snapshot s = snapshotState();
   sample.crank_revs = (uint16_t)s.total_revs;
   
-  // Convert last event time to 1/1024s units
-  // We need to handle the wrap-around of millis() if running for 49 days, 
-  // but for simple casting it might be okay. 
-  // BLE expects a free running timer.
-  sample.crank_evt_1024 = (uint16_t)((s.last_rev * 1024) / 1000);
+  // Convert last event time to 1/1024s units (CPS spec — wraps every 64s, that's fine).
+  // Use 64-bit arithmetic so the intermediate (ms * 1024) doesn't overflow uint32_t
+  // after ~70 minutes of uptime.
+  sample.crank_evt_1024 = (uint16_t)(((uint64_t)s.last_rev * 1024) / 1000);
 
   ready = true;
 }
